Go 1.5 migration
Go 1.4 introduced TLS 1.3 by default, which we can't do since a lot of our clients depend on 1.2.
For example one of 1.3 breaking "features" is inability to explicity set TLS ciphers (in 1.3 they should be picked automatically), and a lot of enterprises depend on it.
In order to make it non breaking change I added new MaxSSLVersion attributes to all needed sections, so by default 1.2 will be used everywhere.

Additionally fixed a bunch of tests (not realted to 1.5 change from what I see).

All the vendoring removed as well!